chore: move deviceEmitter and deviceLogProvider out of appbuilder dir
Move deviceEmitter out of appbuilder dir. Also remove all `companion` related services as they are not required anymore.
Move the device-log-provider out of appbuilder dir. The purpose of the file is to be used instead of the original deviceLogProvider when CLI is used as a library. This provider emits the device logs instead of writing them on the stdout.
To separate the two log providers, rename the one coming from appbuilder dir to `device-log-emitter`. Keep the registration in the `$injector` to be `deviceLogProvider` as it is the only way to keep the current behavior when CLI is used as a library.
